requirements.txt
requests~=2.32.3
typing~=3.7.4.3
dotenv~=0.9.9
python-dotenv~=1.1.0
pytesseract~=0.3.13
googletrans~=4.0.2
pillow~=11.2.1
SpeechRecognition~=3.14.2
gTTS~=2.5.4
pydub~=0.25.1
uvicorn~=0.34.2
fastapi~=0.115.12
opencv-python~=4.11.0.86
numpy~=2.2.5
pydantic~=2.11.4
tensorflow~=2.19.0
mediapipe~=0.10.21
protobuf~=4.25.7
SQLAlchemy~=2.0.40
starlette~=0.46.2